requirements/requirements.txt
#
# This file is autogenerated by hatch-pip-compile with Python 3.12
#
# - aiosql>=9.4
# - click>=8.1.7
# - duckdb>=0.9.2
# - google-cloud-network-management
# - greenlet; sys_platform == "darwin"
# - jinja2>=3.1.3
# - msgspec
# - polars[pyarrow]>=0.20.0
# - rich-click>=1.7.3
# - sqlalchemy>=2.0.25
# - typing-extensions>=4.0.0
#

aiosql==10.1
    # via hatch.envs.default
cachetools==5.3.3
    # via google-auth
certifi==2024.2.2
    # via requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   hatch.envs.default
    #   rich-click
duckdb==0.10.2
    # via hatch.envs.default
google-api-core==2.19.0
    # via google-cloud-network-management
google-auth==2.29.0
    # via
    #   google-api-core
    #   google-cloud-network-management
google-cloud-network-management==1.17.1
    # via hatch.envs.default
googleapis-common-protos==1.63.0
    # via
    #   google-api-core
    #   grpc-google-iam-v1
    #   grpcio-status
greenlet==3.0.3
    # via sqlalchemy
grpc-google-iam-v1==0.13.0
    # via google-cloud-network-management
grpcio==1.63.0
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
grpcio-status==1.62.2
    # via google-api-core
idna==3.7
    # via requests
jinja2==3.1.3
    # via hatch.envs.default
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
msgspec==0.18.6
    # via hatch.envs.default
numpy==1.26.4
    # via pyarrow
polars==0.20.22
    # via hatch.envs.default
proto-plus==1.23.0
    # via
    #   google-api-core
    #   google-cloud-network-management
protobuf==4.25.3
    # via
    #   google-api-core
    #   google-cloud-network-management
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
    #   proto-plus
pyarrow==16.0.0
    # via polars
pyasn1==0.6.0
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.0
    # via google-auth
pygments==2.17.2
    # via rich
requests==2.31.0
    # via google-api-core
rich==13.7.1
    # via rich-click
rich-click==1.7.4
    # via hatch.envs.default
rsa==4.9
    # via google-auth
sqlalchemy==2.0.29
    # via hatch.envs.default
typing-extensions==4.11.0
    # via
    #   hatch.envs.default
    #   rich-click
    #   sqlalchemy
urllib3==2.2.1
    # via requests
